feat: Add Git tag that tracks the Docker edge tag
This commit fixes #18 by adding a Git tag that tracks the Docker edge
tag. This tag is updated after the Docker image is published.

I decided to use a tag for this instead of a branch because we aren't
concerned with history. We only need a reference to the commit used to
build the Docker edge tag.