fix: Move build to the `.devcontainer` directory
The `devcontainers/ci` action will only prebuild the image from the
`.devcontainer` directory, so I consolidated the build files there and
removed the `src` directory.

We now rely on GitHub's Codespaces prebuilds to avoid having to build
the dev container every time a new Codespace is created.

If we want to run a CI build inside the dev container, we can avoid
building the image by using the `devcontainers/ci` action again to pull
the prebuilt image.